1.25.1 to master

+++ app/src/main/java/org/thoughtcrime/securesms/providers/BlobProvider.java
@@ -140,7 +140,12 @@ public class BlobProvider {
    */
   public synchronized void onSessionStart(@NonNull Context context) {
     File directory = getOrCreateCacheDirectory(context, SINGLE_SESSION_DIRECTORY);
-    for (File file : directory.listFiles()) {
+    File[] files = directory.listFiles();
+    if (files == null) {
+      return; 
+    }
+
+    for (File file : files) {
       file.delete();
     }
   }
